Damaging works ish
Player now takes damage and can use functions from MyEnemyBattle

+++ Source/CardLord_HighSchool/MyCharacter.h
@@ -71,5 +71,5 @@ public:
 
 	
 
-	int Health = 100;
+	float Health = 100;
 };
